requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
Pillow==10.2.0
pydantic>=2.4.1,<2.6
pydantic-settings>=2.0.0
loguru==0.7.2
python-dotenv==1.0.1
fal-client>=0.6.0
gspread==6.0.0
google-auth==2.29.0